requirements.txt
fastapi
aioredis
starlette
requests
pytest
uvicorn[standard]
fastapi-cache2[redis]